// client/apps/src/components/commons/forms/SearchField/SearchField.tsx
import styled from 'styled-components'

export const Input = styled.input`
    box-shadow: 1px 1px 1px #B9B7BD;
    overflow: hidden;
    width: 250px;
    height: 50px;
    border: 1px solid #B9B7BD;
    border-radius: 5px;
    outline: none;
    padding: 4px 6px;
    color: gray;
    margin: auto auto;
    &:hover{
    border: 2px solid gray
    }
`